requirements.txt
fastapi
uvicorn
python-dotenv
groq
pytesseract
pillow
streamlit
requests
